+++ src/containers/FoodDetails.jsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
-import { indexOf } from 'lodash-es';
 import CardCarousel from '../components/CardCarousel';
 import blackHeartIcon from '../images/blackHeartIcon.svg';
 import whiteHeartIcon from '../images/whiteHeartIcon.svg';
@@ -11,11 +10,32 @@ const copy = require('clipboard-copy');
 
 const NUMBER_9 = 9;
 
-// const filterIngredients = (recipe) => Object.entries(recipe)
-//   .filter((ingredientIndex) => ingredientIndex[0].startsWith('strIngredient'))
-//   .filter((ingredientIndex) => ingredientIndex[1] !== '')
-//   .filter((ingredientIndex) => ingredientIndex[1] !== null)
-//   .map((ingredientIndex) => ingredientIndex[1]);
+function FavoriteMealRecipe(foods) {
+  const { idMeal, strArea, strCategory,
+    strMeal, strMealThumb } = foods[0];
+  const obj = [{
+    id: idMeal,
+    type: 'comida',
+    area: strArea || '',
+    category: strCategory || '',
+    alcoholicOrNot: '',
+    name: strMeal,
+    image: strMealThumb,
+  }];
+  return obj;
+}
+
+function handleFavorite(checkFavorite, setCheckFavorite, foods) {
+  if (!checkFavorite) {
+    setCheckFavorite(true);
+    localStorage.setItem(
+      'favoriteRecipes', JSON.stringify(FavoriteMealRecipe(foods)),
+    );
+  } else {
+    setCheckFavorite(false);
+    localStorage.removeItem('favoriteRecipes');
+  }
+}
 
 const filterIngredientsAndMeasures = (recipe) => {
   const arrayFromObject = Object.entries(recipe)
@@ -36,65 +56,6 @@ const filterIngredientsAndMeasures = (recipe) => {
   )));
 };
 
-const initLocalStorage = (id) => {
-  const isFavorite = JSON.parse(localStorage.getItem('favoriteRecipes'));
-  console.log(isFavorite);
-  const inProgessRecipes = JSON.parse(localStorage.getItem('inProgressRecipes'));
-  console.log(inProgessRecipes);
-
-  if (inProgessRecipes === null) {
-    localStorage.setItem('inProgressRecipes',
-      JSON.stringify([{ meals: { [id]: [] } }]));
-  }
-  if (isFavorite === null) {
-    localStorage.setItem('favoriteRecipes',
-      JSON.stringify([]));
-  }
-  return { isFavorite, inProgessRecipes };
-};
-
-// const setLocalStorage = (recipe, id) => {
-//   const inProgessRecipes = JSON.parse(localStorage.getItem('inProgressRecipes'));
-//   // const isFavorite = JSON.parse(localStorage.getItem('favoriteRecipes'));
-//   const { cocktails, meals } = inProgessRecipes;
-//   const ingredients = filterIngredients(recipe);
-//   if (inProgessRecipes) {
-//     localStorage.setItem('inProgressRecipes',
-//       JSON.stringify({
-//         cocktails,
-//         meals: { ...meals,
-//           [id]: ingredients } }));
-//   }
-//   localStorage.setItem('inProgressRecipes',
-//     JSON.stringify({
-//       cocktails: {},
-//       meals: {
-//         ...meals, [recipe.idMeal]: ingredients } }));
-// };
-
-const handleFavorite = (recipe, iFavorite) => {
-  console.log('entrou no favorite useEffect');
-  if (iFavorite) {
-    console.log('entrou no isFavorite');
-    const favoriteArray = JSON.parse(localStorage.getItem('favoriteRecipes'));
-    return favoriteArray.length <= 1 ? localStorage.setItem('favoriteRecipes',
-      JSON.stringify([recipe])) : localStorage.setItem('favoriteRecipes',
-      JSON.stringify([...favoriteArray, recipe]));
-  }
-  console.log('saiu do isFavorite');
-
-  const favoriteArray = JSON.parse(localStorage.getItem('favoriteRecipes'));
-  console.log(favoriteArray);
-  if (favoriteArray !== null && favoriteArray.length > 1) {
-    console.log('entrou no delete favorite length maior que 1');
-    return localStorage.setItem('favoriteRecipes', JSON.stringify([
-      favoriteArray.slice(0, favoriteArray[indexOf(recipe[0])]),
-      favoriteArray.slice(favoriteArray[indexOf(recipe[0])]),
-    ]));
-  }
-  // localStorage.setItem('favoriteRecipes', JSON.stringify([]));
-};
-
 const FoodDetails = () => {
   const history = useHistory();
   const [food, setFood] = useState([]);
@@ -102,18 +63,20 @@ const FoodDetails = () => {
   const [loading, setLoading] = useState(true);
   const [start, setStart] = useState(false);
   const [copied, setCopy] = useState(false);
-  const [favorite, setFavorite] = useState(false);
+  const [checkFavorite, setCheckFavorite] = useState(() => {
+    const favorite = JSON.parse(localStorage.getItem('favoriteRecipes'));
+    if (favorite) {
+      return true;
+    }
+    return null;
+  });
   const id = history.location.pathname.slice(NUMBER_9);
+
   useEffect(() => {
     api.fetchMealById(id)
       .then((response) => response.json()).then((result) => setFood(result.meals));
     api.fetchDrinks()
       .then((response) => response.json()).then((result) => setDrinks(result.drinks));
-    console.log(initLocalStorage(id));
-    const { isFavorite } = initLocalStorage(id);
-    if (isFavorite.length > 0) {
-      setFavorite(true);
-    }
   }, [id]);
 
   useEffect(() => {
@@ -125,14 +88,9 @@ const FoodDetails = () => {
 
   useEffect(() => {
     if (start) {
-      // setLocalStorage(food[0]);
       return history.push(`/comidas/${food[0].idMeal}/in-progress`);
     }
   }, [start, food, history]);
-
-  useEffect(() => {
-    handleFavorite(food, favorite);
-  }, [favorite, food]);
 
   const copyToClipBoard = (url) => copy(`http://localhost:3000${url}`)
     .then(() => {
@@ -182,10 +140,12 @@ const FoodDetails = () => {
                   className="share-favorite-buttons"
                   type="submit"
                   data-testid="favorite-btn"
-                  src={ favorite ? blackHeartIcon : whiteHeartIcon }
-                  onClick={ () => setFavorite(!favorite) }
+                  src={ checkFavorite ? blackHeartIcon : whiteHeartIcon }
+                  onClick={ () => handleFavorite(
+                    checkFavorite, setCheckFavorite, food,
+                  ) }
                 >
-                  {favorite
+                  {checkFavorite
                     ? (
                       <img
                         className="rocksGlass"
@@ -232,8 +192,8 @@ const FoodDetails = () => {
               className="btnz2 btn btn-primary start-recipe-btn"
               onClick={ () => setStart(true) }
             >
-              {inProgessRecipes.meals !== {} ? 'Continuar Receita' : 'Iniciar Receita'}
-              {/* Req 39 e 40 passam mas precisamos melhorar isso depois. */}
+              {inProgessRecipes
+                ? 'Continuar Receita' : 'Iniciar Receita'}
             </button>
           </div>
         )}
